Calculate and plot a correlation matrix.

Plots coefficients from a model, along with standard errors. Coefficients that are statistically significant at agiven level are highlighted in darker text. Inspired by https://github.com/jaredlander/coefplot/blob/master/R/coefplot.r but works w/ ggplot2 version > 2.2

plot_heatmap(df, fill_var = "value", remove_same = TRUE,
  remove_half = TRUE, label_pct = TRUE, stroke_colour = "white",
  stroke_size = 0.5, square_tiles = TRUE, fill_palette = brewer.pal(11,
  "RdYlBu"), fill_limits = c(-1, 1), font_normal = "Lato",
  font_semi = "Lato Light", font_light = "Lato Light")

Arguments

df
data frame in a tidy format

Examples

plot_corr(mtcars)

#> Error in plot_corr(mtcars): could not find function "plot_corr"